separate infix tests and add some tests
